Support indenting to align method args, refs #94.
This adds an :argument_alignment option to the indentation_spaces ruler,
allowing you to specify that method declarations and calls should indent
to the first argument on subsequent lines.

You can enable argument alignment by passing it as an option to
indentation spaces in your config:

  style.indentation_spaces 2, level: :error, argument_alignment: true

Currently this only understands method declarations and calls that use
parentheses.